+++ apps/connect/source/ftrack_connect/ui/application.py
@@ -1,6 +1,6 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2014-2023 ftrack
-
+import copy
 import os
 import platform
 import sys
@@ -10,7 +10,6 @@ import uuid
 import logging
 import weakref
 from operator import itemgetter
-import platformdirs
 import time
 import qtawesome as qta
 
@@ -30,7 +29,13 @@ from ftrack_connect import load_icons
 import ftrack_connect
 import ftrack_connect.event_hub_thread as _event_hub_thread
 import ftrack_connect.error
-import ftrack_connect.util
+from ftrack_connect.util import (
+    get_default_plugin_directory,
+    get_plugins_from_path,
+    get_plugin_data,
+    open_directory,
+    PLUGIN_DIRECTORIES,
+)
 import ftrack_connect.ui.theme
 import ftrack_connect.ui.widget.overlay
 from ftrack_connect.ui.widget import uncaught_error as _uncaught_error
@@ -90,6 +95,9 @@ class ConnectWidget(QtWidgets.QWidget):
     #: Signal to emit to request connect to restart.
     requestConnectRestart = QtCore.Signal()
 
+    #: Signal to fetch discovered plugins
+    fetchPlugins = QtCore.Signal(object)
+
     @property
     def session(self):
         '''Return current session.'''
@@ -108,6 +116,17 @@ class ConnectWidget(QtWidgets.QWidget):
         '''Return identifier for widget.'''
         return self.getName().lower().replace(' ', '.')
 
+    def refresh(self):
+        '''Refresh/rebuild widget, called when widget has been fully set up. Should be overridden'''
+        pass
+
+    def get_debug_information(self):
+        '''Return debug information for about dialog, should be overriden.'''
+        return {
+            'name': self.getName(),
+            'identifier': self.getIdentifier(),
+        }
+
 
 class Application(QtWidgets.QMainWindow):
     '''Main application window for ftrack connect.'''
@@ -122,7 +141,68 @@ class Application(QtWidgets.QMainWindow):
     loginSignal = QtCore.Signal(object, object, object)
     loginSuccessSignal = QtCore.Signal()
 
-    _builtin_plugins = []
+    _builtin_widget_plugins = []
+    _widget_plugin_instances = {}
+
+    @property
+    def session(self):
+        '''Return current session.'''
+        return self._session
+
+    @property
+    def plugins(self):
+        return self._plugins
+
+    def __init__(self, theme='system', instance=None, log_level=None):
+        '''Initialise the main application window with *theme*, singleton
+        *instance* and custom *log_level*.'''
+        super(Application, self).__init__()
+        self.logger = logging.getLogger(
+            __name__ + '.' + self.__class__.__name__
+        )
+        self._instance = instance
+        self._session = None
+        self.__connect_start_time = time.time()
+        self._log_level = log_level
+
+        self._create_default_plugin_directory()
+
+        self._discovered_plugins = (
+            self._discover_plugin_data_from_plugin_directories()
+        )
+        self._plugins = self._available_plugin_data_from_plugin_directories()
+
+        # Register widget for error handling.
+        self.uncaughtError = _uncaught_error.UncaughtError(parent=self)
+
+        self._login_server_thread = None
+        self.tray = None
+        if not QtWidgets.QSystemTrayIcon.isSystemTrayAvailable():
+            QtWidgets.QMessageBox.warning(
+                self,
+                'Connect',
+                'No system tray located.\n\nHint: On '
+                'Linux CentOS consider installing '
+                'gnome-shell-extension-top-icons',
+            )
+        else:
+            self._initialise_tray()
+        self._initialise_menu_bar()
+
+        self._set_theme(theme)
+
+        self._application_launcher = None
+
+        self.setObjectName('ftrack-connect-window')
+        self.setWindowTitle('ftrack Connect')
+        self.resize(450, 700)
+        self.move(50, 50)
+
+        self._login_overlay = None
+        self._login_widget = _login.Login(theme=self.theme())
+        self.loginSignal.connect(self.loginWithCredentials)
+        self.loginSuccessSignal.connect(self._post_login_settings)
+        self.login()
 
     def restart(self):
         '''restart connect application'''
@@ -217,62 +297,6 @@ class Application(QtWidgets.QMainWindow):
         if usage_tracker:
             usage_tracker.track('USED-CONNECT', metadata)
 
-    @property
-    def session(self):
-        '''Return current session.'''
-        return self._session
-
-    def __init__(self, theme='system', instance=None, log_level=None):
-        '''Initialise the main application window with *theme*, singleton
-        *instance* and custom *log_level*.'''
-        super(Application, self).__init__()
-        self.logger = logging.getLogger(
-            __name__ + '.' + self.__class__.__name__
-        )
-        self._instance = instance
-        self._session = None
-        self.__connect_start_time = time.time()
-        self._log_level = log_level
-
-        self.defaultPluginDirectory = platformdirs.user_data_dir(
-            'ftrack-connect-plugins', 'ftrack'
-        )
-        self._createDefaultPluginDirectory()
-
-        self.plugin_paths = self._discover_plugin_paths()
-
-        # Register widget for error handling.
-        self.uncaughtError = _uncaught_error.UncaughtError(parent=self)
-
-        self._login_server_thread = None
-        self.tray = None
-        if not QtWidgets.QSystemTrayIcon.isSystemTrayAvailable():
-            QtWidgets.QMessageBox.warning(
-                self,
-                'Connect',
-                'No system tray located.\n\nHint: On '
-                'Linux CentOS consider installing '
-                'gnome-shell-extension-top-icons',
-            )
-        else:
-            self._initialiseTray()
-        self._initialiseMenuBar()
-
-        self.setTheme(theme)
-
-        self.plugins = {}
-
-        self.setObjectName('ftrack-connect-window')
-        self.setWindowTitle('ftrack Connect')
-        self.resize(450, 700)
-        self.move(50, 50)
-
-        self._login_overlay = None
-        self.loginWidget = _login.Login(theme=self.theme())
-        self.loginSignal.connect(self.loginWithCredentials)
-        self.loginSuccessSignal.connect(self._post_login_settings)
-        self.login()
-
     def _post_login_settings(self):
         if self.tray:
             self.tray.show()
@@ -285,7 +309,7 @@ class Application(QtWidgets.QMainWindow):
         '''Return current theme.'''
         return self._theme
 
-    def setTheme(self, theme='system'):
+    def _set_theme(self, theme='system'):
         '''Set *theme*.'''
         if theme not in ['light', 'dark']:
             theme = self.system_theme()
@@ -317,7 +341,7 @@ class Application(QtWidgets.QMainWindow):
         if event['topic'] != 'ftrack.connect':
             return
 
-        self._routeEvent(event)
+        self._route_event(event)
 
     def logout(self):
         '''Clear stored credentials and quit Connect.'''
@@ -400,7 +424,7 @@ class Application(QtWidgets.QMainWindow):
     def login(self):
         '''Login using stored credentials or ask user for them.'''
         credentials = self._get_credentials()
-        self.showLoginWidget()
+        self._show_login_widget()
 
         if credentials:
             # Try to login.
@@ -410,23 +434,23 @@ class Application(QtWidgets.QMainWindow):
                 credentials['api_key'],
             )
 
-    def showLoginWidget(self):
+    def _show_login_widget(self):
         '''Show the login widget.'''
         self._login_overlay = ftrack_connect.ui.widget.overlay.CancelOverlay(
-            self.loginWidget, message='<h2>Signing in<h2/>'
+            self._login_widget, message='<h2>Signing in<h2/>'
         )
 
         self._login_overlay.hide()
-        self.setCentralWidget(self.loginWidget)
-        self.loginWidget.login.connect(self._login_overlay.show)
-        self.loginWidget.login.connect(self.loginWithCredentials)
-        self.loginError.connect(self.loginWidget.loginError.emit)
+        self.setCentralWidget(self._login_widget)
+        self._login_widget.login.connect(self._login_overlay.show)
+        self._login_widget.login.connect(self.loginWithCredentials)
+        self.loginError.connect(self._login_widget.loginError.emit)
         self.loginError.connect(self._login_overlay.hide)
         self.focus()
 
         # Set focus on the login widget to remove any focus from its child
         # widgets.
-        self.loginWidget.setFocus()
+        self._login_widget.setFocus()
         self._login_overlay.hide()
 
     def _get_api_plugin_paths(self):
@@ -438,8 +462,18 @@ class Application(QtWidgets.QMainWindow):
             if apiPluginPath and apiPluginPath not in api_plugin_paths:
                 api_plugin_paths.append(os.path.expandvars(apiPluginPath))
 
-        for connect_plugin_path in self.plugin_paths:
-            api_plugin_paths.append(os.path.join(connect_plugin_path, 'hook'))
+        checked_plugins = []
+        for plugin in self.plugins:
+            # Always pick the
+            if plugin['name'] in checked_plugins:
+                continue
+            if plugin['incompatible']:
+                self.logger.warning(
+                    f'Ignoring plugin that is incompatible: {plugin["path"]}'
+                )
+                continue
+            checked_plugins.append(plugin['name'])
+            api_plugin_paths.append(os.path.join(plugin['path'], 'hook'))
 
         return api_plugin_paths
 
@@ -586,7 +620,7 @@ class Application(QtWidgets.QMainWindow):
                     self.session
                 )
                 scenario_widget.configuration_completed.connect(
-                    self.location_configuration_finished
+                    self._location_configuration_finished
                 )
                 self.setCentralWidget(scenario_widget)
                 self.focus()
@@ -594,10 +628,10 @@ class Application(QtWidgets.QMainWindow):
                 return
 
         # No change so build if needed
-        self.location_configuration_finished(reconfigure_session=False)
+        self._location_configuration_finished(reconfigure_session=False)
         self.loginSuccessSignal.emit()
 
-    def location_configuration_finished(self, reconfigure_session=True):
+    def _location_configuration_finished(self, reconfigure_session=True):
         '''Continue connect setup after location configuration is done.'''
         if reconfigure_session:
             try:
@@ -610,7 +644,7 @@ class Application(QtWidgets.QMainWindow):
         self._session = self._setup_session()
 
         try:
-            self.configureConnectAndDiscoverPlugins()
+            self._configure_connect_and_discover_plugins()
         except Exception as error:
             self.logger.exception(u'Error during location configuration:')
             self._report_session_setup_error(error)
@@ -635,7 +669,7 @@ class Application(QtWidgets.QMainWindow):
             msgBox.setText('\n\n'.join(problems))
             msgBox.exec_()
 
-    def configureConnectAndDiscoverPlugins(self):
+    def _configure_connect_and_discover_plugins(self):
         '''Configure connect and load plugins.'''
 
         self.tabPanel = _tab_widget.TabWidget()
@@ -646,7 +680,7 @@ class Application(QtWidgets.QMainWindow):
             'topic=ftrack.connect and source.user.username="{0}"'.format(
                 self.session.api_user
             ),
-            self._relayEventHubEvent,
+            self._relay_event,
         )
         self.eventHubSignal.connect(self._onConnectTopicEvent)
 
@@ -670,72 +704,112 @@ class Application(QtWidgets.QMainWindow):
 
         self._discover_connect_widgets()
 
-    def _discover_plugin_paths(self):
-        '''Return a list of paths to pass to ftrack_api.Session()'''
+    def _gather_plugins(self, plugin_directory, source_index=None):
+        '''Return plugin data from *plugin_directory*.'''
 
-        plugin_paths = []
-
-        plugin_paths.extend(self._gather_plugins(self.defaultPluginDirectory))
-
-        for connectPluginPath in os.environ.get(
-            'FTRACK_CONNECT_PLUGIN_PATH', ''
-        ).split(os.pathsep):
-            plugin_paths.extend(
-                self._gather_plugins(os.path.expandvars(connectPluginPath))
-            )
-
-        plugin_paths = list(set(plugin_paths))
-
-        self.logger.info(
-            u'Connect plugin directories: {0}'.format(', '.join(plugin_paths))
+        result = []
+        if not plugin_directory:
+            return result
+        self.logger.debug(
+            u'Searching {0!r} for plugin hooks.'.format(plugin_directory)
         )
-
-        return plugin_paths
-
-    def _gather_plugins(self, path):
-        '''Return plugin hooks from *path*.'''
-
-        paths = []
-        if not path:
-            return paths
-        self.logger.debug(u'Searching {0!r} for plugin hooks.'.format(path))
-        if os.path.isdir(path):
-            for candidate in ftrack_connect.util.get_plugins_from_path(path):
-                candidate_path = os.path.join(path, candidate)
-                if os.path.isdir(candidate_path):
-                    if ftrack_connect.util.is_incompatible_plugin(
-                        ftrack_connect.util.get_plugin_data(candidate_path)
-                    ):
-                        self.logger.warning(
-                            f'Ignoring plugin that is incompatible: {candidate_path}'
-                        )
-                        continue
-                    full_hook_path = os.path.join(candidate_path, 'hook')
-                    if (
-                        os.path.isdir(full_hook_path)
-                        and candidate_path not in paths
-                    ):
-                        paths.append(candidate_path)
+        if os.path.isdir(plugin_directory):
+            for candidate in get_plugins_from_path(plugin_directory):
+                candidate_path = os.path.join(plugin_directory, candidate)
+                plugin_data = get_plugin_data(candidate_path)
+                if source_index >= 0:
+                    plugin_data['source_index'] = source_index
+                result.append(plugin_data)
 
         self.logger.debug(
-            u'Found {0!r} plugin hooks in {1!r}.'.format(paths, path)
+            u'Found {0!r} plugin hooks in {1!r}.'.format(
+                result, plugin_directory
+            )
         )
 
-        return paths
+        return result
 
-    def _relayEventHubEvent(self, event):
+    def _available_plugin_data_from_plugin_directories(self):
+        '''Return a list of plugin data'''
+
+        result = []
+        i = 0
+        for plugin_base_directory in PLUGIN_DIRECTORIES:
+            current_dir_plugins = []
+            for plugin in self._gather_plugins(
+                plugin_base_directory, source_index=i
+            ):
+                # Append plugin if not already in more prioritized paths. - top plugin path takes preference
+                found = False
+                for existing_plugin in result:
+                    if (
+                        existing_plugin['name'].lower()
+                        == plugin['name'].lower()
+                    ):
+                        found = True
+                        break
+                # Make sure we pick the compatible latest version of the plugin from the current folder
+                for current_dir_plugin in current_dir_plugins:
+                    if (
+                        current_dir_plugin['name'].lower()
+                        == plugin['name'].lower()
+                    ):
+                        if (
+                            current_dir_plugin['incompatible']
+                            and not plugin['incompatible']
+                        ):
+                            current_dir_plugins.remove(current_dir_plugin)
+                            break
+                        elif (
+                            current_dir_plugin['deprecated']
+                            and not plugin['incompatible']
+                            and not plugin['deprecated']
+                        ):
+                            current_dir_plugins.remove(current_dir_plugin)
+                            break
+                        elif (
+                            not plugin['incompatible']
+                            and not plugin['deprecated']
+                        ):
+                            if (
+                                current_dir_plugin['version']
+                                < plugin['version']
+                            ):
+                                current_dir_plugins.remove(current_dir_plugin)
+                                break
+                        found = True
+                        break
+                if not found:
+                    current_dir_plugins.append(plugin)
+            result.extend(current_dir_plugins)
+            i += 1
+        return result
+
+    def _discover_plugin_data_from_plugin_directories(self):
+        '''Return a list of all plugin data from the plugin directories'''
+
+        result = []
+        i = 0
+        for plugin_base_directory in PLUGIN_DIRECTORIES:
+            result.extend(
+                self._gather_plugins(plugin_base_directory, source_index=i)
+            )
+            i += 1
+        return result
+
+    def _relay_event(self, event):
         '''Relay all ftrack.connect events.'''
         self.eventHubSignal.emit(event)
 
-    def _initialiseTray(self):
+    def _initialise_tray(self):
         '''Initialise and add application icon to system tray.'''
-        self.trayMenu = self._createTrayMenu()
+        self.trayMenu = self._create_tray_menu()
 
         self.tray = QtWidgets.QSystemTrayIcon(self)
 
         self.tray.setContextMenu(self.trayMenu)
 
-    def _initialiseMenuBar(self):
+    def _initialise_menu_bar(self):
         '''Initialise and add connect widget to widgets menu.'''
 
         self.menu_bar = QtWidgets.QMenuBar()
@@ -744,7 +818,7 @@ class Application(QtWidgets.QMainWindow):
         self.menu_widget = widget_menu
         self.menu_bar.setVisible(False)
 
-    def _createTrayMenu(self):
+    def _create_tray_menu(self):
         '''Return a menu for system tray.'''
         menu = QtWidgets.QMenu(self)
 
@@ -761,11 +835,11 @@ class Application(QtWidgets.QMainWindow):
         openPluginDirectoryAction = QtWidgets.QAction(
             'Open plugin directory',
             self,
-            triggered=self.openDefaultPluginDirectory,
+            triggered=self._open_default_plugin_directory,
         )
 
         aboutAction = QtWidgets.QAction(
-            'About', self, triggered=self.showAbout
+            'About', self, triggered=self._show_about
         )
 
         alwaysOnTopAction = QtWidgets.QAction('Always on top', self)
@@ -773,7 +847,7 @@ class Application(QtWidgets.QMainWindow):
             'Restart', self, triggered=self.restart
         )
         alwaysOnTopAction.setCheckable(True)
-        alwaysOnTopAction.triggered[bool].connect(self.setAlwaysOnTop)
+        alwaysOnTopAction.triggered[bool].connect(self._set_always_on_top)
 
         menu.addAction(aboutAction)
         menu.addSeparator()
@@ -795,15 +869,12 @@ class Application(QtWidgets.QMainWindow):
         '''Find and load connect widgets in search paths.'''
 
         event = ftrack_api.event.base.Event(topic=ConnectWidgetPlugin.topic)
-        disable_startup_widget = bool(
-            os.getenv('FTRACK_CONNECT_DISABLE_STARTUP_WIDGET', False)
-        )
         responses = self.session.event_hub.publish(event, synchronous=True)
 
         # Load icons
         load_icons(os.path.join(os.path.dirname(__file__), '..', 'fonts'))
 
-        for plugin_class in self._builtin_plugins + responses:
+        for plugin_class in self._builtin_widget_plugins + responses:
             widget_plugin = None
             try:
                 widget_plugin = plugin_class(self.session)
@@ -823,15 +894,15 @@ class Application(QtWidgets.QMainWindow):
                 continue
             try:
                 identifier = widget_plugin.getIdentifier()
-                if not self.plugins.get(identifier):
-                    self.plugins[identifier] = widget_plugin
+                if not self._widget_plugin_instances.get(identifier):
+                    self._widget_plugin_instances[identifier] = widget_plugin
                 else:
                     self.logger.debug(
                         'Widget {} already registered'.format(identifier)
                     )
                     continue
 
-                self.addPlugin(widget_plugin)
+                self._add_plugin(widget_plugin)
 
             except Exception as error:
                 self.logger.warning(
@@ -840,7 +911,7 @@ class Application(QtWidgets.QMainWindow):
                     )
                 )
 
-    def _routeEvent(self, event):
+    def _route_event(self, event):
         '''Route websocket *event* to publisher plugin.
 
         Expect event['data'] to contain:
@@ -856,7 +927,7 @@ class Application(QtWidgets.QMainWindow):
         action = event['data']['action']
 
         try:
-            pluginInstance = self.plugins[plugin]
+            pluginInstance = self._widget_plugin_instances[plugin]
         except KeyError:
             raise ftrack_connect.error.ConnectError(
                 'Plugin "{0}" not found.'.format(plugin)
@@ -873,16 +944,20 @@ class Application(QtWidgets.QMainWindow):
 
         method(event)
 
-    def _onWidgetRequestApplicationFocus(self, widget):
+    def _on_fetch_plugins_callback(self, callback_fn):
+        '''Call *callback_fn* with list of active Connect plugins'''
+        callback_fn(self.plugins)
+
+    def _on_widget_request_application_focus_callback(self, widget):
         '''Switch tab to *widget* and bring application to front.'''
         self.tabPanel.setCurrentWidget(widget)
         self.focus()
 
-    def _onWidgetRequestApplicationClose(self, widget):
+    def _on_widget_request_acpplication_close_callback(self, widget):
         '''Hide application upon *widget* request.'''
         self.hide()
 
-    def addPlugin(self, plugin, name=None):
+    def _add_plugin(self, plugin, name=None):
         '''Add *plugin* in new tab with *name* and *identifier*.
 
         *plugin* should be an instance of :py:class:`ApplicationPlugin`.
@@ -899,30 +974,33 @@ class Application(QtWidgets.QMainWindow):
 
         try:
             icon = qta.icon(plugin.icon)
-        except Exception as err:
+        except Exception:
             icon = QtGui.QIcon(plugin.icon)
 
         self.tabPanel.addTab(plugin, icon, name)
 
         # Connect standard plugin events.
+        plugin.fetchPlugins.connect(self._on_fetch_plugins_callback)
         plugin.requestApplicationFocus.connect(
-            self._onWidgetRequestApplicationFocus
+            self._on_widget_request_application_focus_callback
         )
         plugin.requestApplicationClose.connect(
-            self._onWidgetRequestApplicationClose
+            self._on_widget_request_acpplication_close_callback
         )
         plugin.requestConnectRestart.connect(self.restart)
 
+        plugin.refresh()
+
         self.logger.debug(f'Plugin {name}({plugin.__class__.__name__}) added')
 
-    def removePlugin(self, plugin):
+    def _remove_plugin(self, plugin):
         '''Remove plugin registered with *identifier*.
 
         Raise :py:exc:`KeyError` if no plugin with *identifier* has been added.
 
         '''
         identifier = plugin.getIdentifier()
-        registered_plugin = self.plugins.get(identifier)
+        registered_plugin = self._widget_plugin_instances.get(identifier)
 
         if registered_plugin is None:
             self.logger.warning(
@@ -941,7 +1019,7 @@ class Application(QtWidgets.QMainWindow):
         self.show()
         self.raise_()
 
-    def setAlwaysOnTop(self, state):
+    def _set_always_on_top(self, state):
         '''Set the application window to be on top'''
         if state:
             self.setWindowFlags(
@@ -953,61 +1031,109 @@ class Application(QtWidgets.QMainWindow):
             )
         self.focus()
 
-    def showAbout(self):
+    def _show_about(self):
         '''Display window with about information.'''
+        from ftrack_connect.plugin_manager import PluginManager
+        from ftrack_connect.plugin_manager.processor import ROLES, STATUSES
+
         self.focus()
 
         aboutDialog = _about.AboutDialog(self)
 
-        environmentData = os.environ.copy()
-        environmentData.update(
+        environment_data = os.environ.copy()
+        environment_data.update(
             {
                 'PLATFORM': platform.platform(),
                 'PYTHON_VERSION': platform.python_version(),
             }
         )
 
-        versionData = [
-            {
-                'name': 'ftrack connect',
-                'version': ftrack_connect.__version__,
-                'core': True,
-                'debug_information': environmentData,
-            }
-        ]
+        debug_information = {'environment': environment_data, 'widgets': []}
+
+        # Provide debug information from application launcher
+        if self._application_launcher:
+            debug_information[
+                'application_launcher'
+            ] = self._application_launcher.get_debug_information()
+
+        # Provide information from builtin widget plugins.
+        for plugin in self._widget_plugin_instances:
+            if isinstance(plugin, ConnectWidget):
+                debug_information['widgets'].append(
+                    plugin.get_debug_information()
+                )
+
+        connect_version_data = {
+            'name': 'ftrack connect',
+            'version': ftrack_connect.__version__,
+            'core': True,
+            'debug_information': debug_information,
+        }
+
+        result = [connect_version_data]
+
+        # Append all available plugins
+        for plugin in self._discovered_plugins:
+            result.append(copy.deepcopy(plugin))
 
         # Gather information about API versions and other
-        # plugin hooks.
+        # plugin hooks through event, append or add to result
 
         try:
             event = ftrack_api.event.base.Event(
                 topic='ftrack.connect.plugin.debug-information'
             )
 
-            responses = self.session.event_hub.publish(event, synchronous=True)
+            raw_responses = self.session.event_hub.publish(
+                event, synchronous=True
+            )
 
-            for response in responses:
+            responses = []
+            for response in raw_responses:
                 if isinstance(response, dict):
-                    versionData.append(response)
+                    responses.append(response)
                 elif isinstance(response, list):
-                    versionData = versionData + response
-
+                    responses = response
+            for response in responses:
+                found = False
+                for plugin_data in result:
+                    if plugin_data['name'] == response['name']:
+                        found = True
+                        # Update information
+                        plugin_data.update(response)
+                        break
+                if not found:
+                    result.append(response)
         except Exception as error:
             self.logger.error(error)
 
-        sorted_version_data = sorted(versionData, key=itemgetter('name'))
+        current_plugins_path = [plugin['path'] for plugin in self.plugins]
+        # Highlight compatibility and source of plugins
+        for plugin_data in result:
+            if plugin_data.get('core'):
+                continue
+            tags = []
+            if plugin_data.get('incompatible'):
+                tags.append('Incompatible')
+            elif plugin_data.get('deprecated'):
+                tags.append('Deprecated')
+            if plugin_data['path'] not in current_plugins_path:
+                tags.append('Ignored')
+            plugin_data['tags'] = ",".join(tags)
+
+        sorted_version_data = sorted(result, key=itemgetter('name'))
 
         aboutDialog.setInformation(
             versionData=sorted_version_data,
             server=os.environ.get('FTRACK_SERVER', 'Not set'),
             user=self.session.api_user,
-            widget_plugins=self.plugins,
+            widget_plugins=self._widget_plugin_instances,
         )
 
         aboutDialog.exec_()
 
-    def _createDefaultPluginDirectory(self):
-        directory = self.defaultPluginDirectory
+    def _create_default_plugin_directory(self):
+        directory = get_default_plugin_directory()
 
         if not os.path.exists(directory):
             # Create directory if not existing.
@@ -1018,22 +1144,22 @@ class Application(QtWidgets.QMainWindow):
 
         return directory
 
-    def openDefaultPluginDirectory(self):
+    def _open_default_plugin_directory(self):
         '''Open default plugin directory in platform default file browser.'''
 
         try:
-            self._createDefaultPluginDirectory()
+            self._create_default_plugin_directory()
         except OSError:
             messageBox = QtWidgets.QMessageBox(parent=self)
             messageBox.setIcon(QtWidgets.QMessageBox.Warning)
             messageBox.setText(
                 u'Could not open or create default plugin '
-                u'directory: {0}.'.format(self.defaultPluginDirectory)
+                u'directory: {0}.'.format(get_default_plugin_directory())
             )
             messageBox.exec_()
             return
 
-        ftrack_connect.util.open_directory(self.defaultPluginDirectory)
+        open_directory(get_default_plugin_directory())
 
     def _discover_applications(self):
         '''Walk through Connect plugins and pick up application launcher
@@ -1042,16 +1168,16 @@ class Application(QtWidgets.QMainWindow):
 
         self.logger.debug('Discovering applications launcher configs.')
 
-        for connect_plugin_path in self.plugin_paths:
+        for connect_plugin_path in [plugin['path'] for plugin in self.plugins]:
             launcher_config_path = os.path.join(connect_plugin_path, 'launch')
             if os.path.isdir(launcher_config_path):
                 launcher_config_paths.append(launcher_config_path)
 
         # Create store containing launchable applications.
-        applications = DiscoverApplications(
+        self._application_launcher = DiscoverApplications(
             self.session, launcher_config_paths
         )
-        applications.register()
+        self._application_launcher.register()
 
     def _configure_action_launcher_widget(self):
         '''Append action launcher widget to list of build in
@@ -1060,7 +1186,7 @@ class Application(QtWidgets.QMainWindow):
         from ftrack_connect.action_launcher import ActionLauncherWidget
 
         # Add together with discovered widgets
-        self._builtin_plugins.append(ActionLauncherWidget)
+        self._builtin_widget_plugins.append(ActionLauncherWidget)
 
     def _configure_plugin_manager_widget(self):
         '''Append plugin manager widget to list of build in
@@ -1069,7 +1195,7 @@ class Application(QtWidgets.QMainWindow):
         from ftrack_connect.plugin_manager import PluginManager
 
         # Add together with discovered widgets
-        self._builtin_plugins.append(PluginManager)
+        self._builtin_widget_plugins.append(PluginManager)
 
     def closeEvent(self, event):
         ''' ' Quit application when main window is closed, and no tray'''