+++ projects/connect-timetracker-widget/release_notes.md
@@ -1,8 +1,9 @@
 # ftrack Connect Timetracker widget release Notes
 
 ## v24.2.0
-2024-02-13
+2024-02-14
 
+* [fix] Fixed bug with wrong plugin name reported back with debug information to Connect about dialog.
 * [changed] Ported from ftrack-connect-timetracker-widget repo and aligned with Connect 3
 * [changed] Use poetry as build system.